+++ docs/changelog.rst
@@ -4,5 +4,6 @@ Changelog
 `Next Release`_
 ---------------
 - Add :func:`coercion.stringify`
+- Add :func:`coercion.normalize_collection`
 
 .. _Next Release: https://github.com/dave-shawley/coercion/compare/0.0.0...master